fix warning for narrowing conversion

+++ src/Modules/IFTMonitor/IFTMonitorModule.hpp
@@ -91,7 +91,7 @@ class IFTMonitorModule : public MonitorBaseModule {
   double rms(double* x, int n);
 
   std::map<int, std::vector<SpacePoint>> m_spacepoints = {};
-  uint32_t m_stationID = 0;
+  uint8_t m_stationID = 0;
   const std::string m_hit_maps[3] = {"hitmap_l0", "hitmap_l1", "hitmap_l2"};
   const std::string m_prefix_hname_hitp = "hitpattern_layer";
   const std::vector<std::string> m_hitp_categories = { "000", "001", "010", "011", "100", "110"};
